services/admin-svc/requirements.txt
# AIVO Admin Service Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
aioredis==2.0.1

# HTTP client
httpx==0.25.2

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.8

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
faker==20.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production server
gunicorn==21.2.0
